requirements.txt
future
pynput
Pillow
DateTime
requests
pygame
PyAutoGUI